Undo rename of shadow blot
