+++ src/registry.ts
@@ -1,6 +1,6 @@
 import Attributor from './attributor/attributor';
 import Formattable from './blot/abstract/formattable';
-import Blot from "./blot/abstract/blot";
+import Blot from "./blot/abstract/shadow";
 
 export type BlotConstructor = typeof Blot;
